remove leaflet map
