--- packages/celer-web-app/experiments.json
@@ -1,12 +1,0 @@
-{
-    "EnhancedScrollTrackerEnabled": false,
-    "MapSyncToDocScrollEnabled": false,
-    "ExportCustomSplits": true,
-    "CleanSplitNames": true,
-    "NoTrackDocPos": true,
-    "WarnNegativeVar": false,
-    "BetterMap": true,
-    "NewASP": true,
-    "EnableDeprecatedRouteBundle": true,
-    "InferCoord": true
-}
