+++ packages/celer-web-app/src/app/AppState.tsx
@@ -3,8 +3,8 @@ import React, { useCallback, useEffect, useMemo, useState } from "react";
 import { BannerType, Compiler, Coord, RouteAssemblySection, SplitType, StringParser } from "core/compiler";
 import { AppExperimentsContext, AppState as ContextState, AppStateContext, useDocument } from "core/context";
 import { RouteEngine } from "core/engine";
-import { useExpNewASP, useExpWarnNegativeVar, useExpEnableDeprecatedRouteBundle, useExpBetterMap, useExpInferCoord, useExpNewDP } from "core/experiments";
-import { InGameCoordinates, MapCore, MapCoreLeaflet, MapEngine, MapIcon, MapLine } from "core/map";
+import { useExpNewASP, useExpWarnNegativeVar, useExpEnableDeprecatedRouteBundle, useExpInferCoord, useExpNewDP } from "core/experiments";
+import { InGameCoordinates, MapEngine, MapIcon, MapLine } from "core/map";
 import { MapDisplayMode, MapDisplayModeStorage, SplitSettingStorage, Theme, ThemeStorage } from "core/settings";
 import { ensureMetadata, addRouteScriptDeprecationMessage, ensureConfig } from "data/bundler";
 import { SourceObject, RouteMetadata, RouteConfig } from "data/libs";
@@ -25,8 +25,6 @@ const compiler = new Compiler();
 const routeEngine = new RouteEngine();
 const mapEngine = new MapEngine();
 
-// Wrapper for L.Map which is deprecated
-const deprecatedMapCore = new MapCoreLeaflet();
 
 export const AppStateProvider: React.FC = ({children})=>{
 	if(useExpNewASP()){
@@ -35,26 +33,6 @@ export const AppStateProvider: React.FC = ({children})=>{
 	return <AppStateProviderOld>{children}</AppStateProviderOld>;
 };
 
-// empty impl for pigeon map
-class DummyMapCore implements MapCore {
-	setMap(_map: Map): void {
-		// do nothing
-	}
-	setIcons(_icons: MapIcon[]): void {
-		// do nothing
-	}
-	setLines(_lines: MapLine[]): void {
-		// do nothing
-	}
-	centerMap(_coord: Coord): void {
-		// do nothing
-	}
-	invalidateSize(): void {
-		// do nothing
-	}
-	
-}
-const dummyMapCore = new DummyMapCore();
 
 export const AppStateProviderFC: React.FC = ({children})=>{
 	const warnNegativeVar = useExpWarnNegativeVar();
@@ -66,7 +44,6 @@ export const AppStateProviderFC: React.FC = ({children})=>{
 		routeEngine.inferCoord = enableInferCoord;
 	}, [warnNegativeVar, enableInferCoord]);
 	const enableDeprecated = useExpEnableDeprecatedRouteBundle();
-	const enablePigeonMap = useExpBetterMap();
 	// set up AppState
 	const [mapDisplayMode, setMapDisplayMode] = useState(
 		()=>MapDisplayModeStorage.load()
@@ -153,10 +130,7 @@ export const AppStateProviderFC: React.FC = ({children})=>{
 	const {docLines, mapIcons, mapLines} = useMemo(()=>{
 		const docLines = routeEngine.compute(routeAssembly);
 		const [mapIcons, mapLines] = mapEngine.compute(docLines);
-		if(!enablePigeonMap){
-			deprecatedMapCore.setIcons(mapIcons);
-			deprecatedMapCore.setLines(mapLines);
-		}
+
 		return {
 			docLines,
 			mapIcons,
@@ -176,7 +150,6 @@ export const AppStateProviderFC: React.FC = ({children})=>{
 	return (
 		<AppStateContext.Provider value={{
 			mapDisplayMode,
-			mapCore: enablePigeonMap ? dummyMapCore : deprecatedMapCore,
 			theme,
 			splitSetting,
 			enableSubsplits,
@@ -221,7 +194,7 @@ export class AppStateProviderOld extends React.Component<EmptyObject, AppState> 
 			theme: ThemeStorage.load(),
 			splitSetting: SplitSettingStorage.load(),
 			enableSubsplits: LocalStorageWrapper.load(ENABLE_SUBSPLITS_KEY, false),
-			mapCore: deprecatedMapCore,
+			
 			docScrollToLine: LocalStorageWrapper.load(DOC_LINE_POS_KEY, 0),
 			docCurrentLine: 0,
 			mapCenter: undefined,
@@ -311,11 +284,6 @@ export class AppStateProviderOld extends React.Component<EmptyObject, AppState> 
 		if(metadata.name){
 			document.title = `${metadata.name} - Celer`;
 		}
-		const PigeonMapEnabled = this.context("BetterMap");
-		if(!PigeonMapEnabled){
-			this.state.mapCore.setIcons(mapIcons);
-			this.state.mapCore.setLines(mapLines);
-		}
 	}
 
 	private setMapCenter(igc: InGameCoordinates) {
@@ -328,7 +296,6 @@ export class AppStateProviderOld extends React.Component<EmptyObject, AppState> 
 
 		return <AppStateContext.Provider value={{
 			mapDisplayMode: this.state.mapDisplayMode,
-			mapCore: this.state.mapCore,
 			theme: this.state.theme,
 			splitSetting: this.state.splitSetting,
 			enableSubsplits: this.state.enableSubsplits,
