+++ packages/celer-web-app/src/data/docs/index.ts
@@ -1,4 +1,3 @@
-export * from "./docline";
 export * from "./functions";
 export * from "./presets";
 export * from "./routescript";
