--- packages/celer-web-app/src/data/libs/scripts.js
@@ -1,2 +1,0 @@
-import "./LeafletTileLayerNoGap.js";
-import "leaflet-arrowheads";
