+++ packages/celer-web-app/src/index.tsx
@@ -1,8 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import { AppMain } from "app/AppMain";
-// Clean up when PigeonMap is enabled
-/*import-validation-exempt*/import "data/libs/scripts";
 import "./index.css";
 
 ReactDOM.render(
